harvester_robot_tests/requirements.txt
# Robot Framework and core libraries
robotframework==7.0
robotframework-requests==0.9.5

# Python libraries
requests==2.31.0
pyyaml==6.0.1
paramiko==3.4.0
bcrypt==4.1.2
cryptography==42.0.4

# SSH and networking
scp==0.14.5
sshpubkeys==3.3.1

# Cloud providers and storage
boto3==1.34.0
botocore==1.34.0

# Utilities
polling2==0.5.0
python-dateutil==2.8.2
jinja2==3.1.3

# Kubernetes client
kubernetes==29.0.0

# Testing utilities
pytest==8.0.0
pytest-dependency==0.6.0

# Logging and monitoring
coloredlogs==15.0.1


# UI automation
playwright==1.56.0